refactor: move to own component
